baichuan_omni_requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.3.5
aiohttp==3.10.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
altair==5.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.2
anyio==4.4.0
argcomplete==3.5.1
argparse==1.4.0
attrs==24.2.0
audioread==3.0.1
av==12.3.0
babel==2.16.0
backcall==0.2.0
bangla==0.0.2
beartype==0.18.5
beautifulsoup4==4.12.3
bleach==6.2.0
blessed==1.20.0
blinker==1.8.2
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
boto3==1.35.67
botocore==1.35.67
braceexpand==0.1.7
brotli==1.1.0
cachetools==5.4.0
cairocffi==1.7.1
cairosvg==2.7.1
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
crcmod==1.7
cryptography==43.0.0
cssselect2==0.7.0
cycler==0.12.1
cymem==2.0.8
cython==3.0.11
dacite==1.7.0
dataclasses-json==0.6.7
datasets==2.21.0
dateparser==1.1.8
datetime==5.5
debugpy==1.8.5
decorator==4.4.2
decord==0.6.0
defusedxml==0.7.1
deprecated==1.2.14
diffusers==0.24.0
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
docopt==0.6.2
easydict==1.13
edit-distance==1.0.6
editdistance==0.8.1
ego4d==1.7.3
einops==0.7.0
einx==0.3.0
english-text-normalization==0.0.3
et-xmlfile==2.0.0
fastapi==0.115.2
fastjsonschema==2.21.1
# fasttext==0.9.2
ffmpeg-python==0.2.0
ffmpy==0.4.0
fiftyone==1.0.2
fiftyone-brain==0.17.0
fiftyone-db==1.1.7
filelock==3.13.1
fire==0.4.0
flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.3.1
future==1.0.0
g2pkk==0.1.2
gdown==5.1.0
gevent==24.2.1
gitdb==4.0.11
gitpython==3.1.43
glob2==0.7
gpustat==1.1.1
gradio==5.11.0
gradio-client==1.5.3
graphql-core==3.2.5
greenlet==3.1.0
grpcio==1.65.4
h11==0.14.0
h2==4.1.0
h5py==3.12.1
halo==0.0.31
hangul-romanize==0.1.0
hjson==3.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface==0.0.1
huggingface-hub==0.25.1
humanfriendly==10.0
humanize==4.11.0
hydra-core==1.3.2
hypercorn==0.17.3
hyperframe==6.0.1
hyperpyyaml==1.2.2
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
imagesize==1.4.1
importlib-metadata==8.4.0
importlib-resources==6.4.0
inflate64==1.0.0
inflect==7.3.1
iopath==0.1.10
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
jinja2==3.1.3
jiter==0.5.0
jmespath==0.10.0
joblib==1.4.2
jq==1.8.0
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.2
jupyter-core==5.7.2
jupyterlab-pygments==0.3.0
kagglehub==0.3.4
kaldiio==2.18.0
kaleido==0.2.1
kiwisolver==1.4.5
langcodes==3.4.0
langdetect==1.0.9
language-data==1.2.0
lazy-loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
log-symbols==0.0.14
lxml==5.3.0
marisa-trie==1.2.0
markdown==3.6
markdown-it-py==3.0.0
markupsafe==2.1.5
marshmallow==3.24.0
matplotlib==3.9.2
mdurl==0.1.2
mistune==3.1.0
modelscope==1.17.1
mongoengine==0.24.2
more-itertools==10.4.0
motor==3.5.3
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
multivolumefile==0.2.3
murmurhash==1.0.10
mutagen==1.47.0
mypy-extensions==1.0.0
narwhals==1.4.1
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8.8
ninja==1.11.1.1
nncore==0.4.5
num2words==0.5.13
numba==0.60.0
numpy==1.26.3
# nvidia-cublas-cu12==12.4.2.65
# nvidia-cuda-cupti-cu12==12.4.99
# nvidia-cuda-nvrtc-cu12==12.4.99
# nvidia-cuda-runtime-cu12==12.4.99
# nvidia-cudnn-cu12==9.1.0.70
# nvidia-cufft-cu12==11.2.0.44
# nvidia-curand-cu12==10.3.5.119
# nvidia-cusolver-cu12==11.6.0.99
# nvidia-cusparse-cu12==12.3.0.142
# nvidia-ml-py==12.555.43
# nvidia-nccl-cu12==2.20.5
# nvidia-nvjitlink-cu12==12.4.99
# nvidia-nvtx-cu12==12.4.99
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.19.2
openai==1.43.1
opencc-python-reimplemented==0.1.7
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parquet-tools==0.2.16
pedalboard==0.9.16
pickleshare==0.7.5
pillow==10.2.0
pillow-simd==9.5.0.post2
pip==24.3.1
pipreqs==0.5.0
platformdirs==4.2.2
plotly==5.24.1
pooch==1.8.2
portalocker==2.0.0
pprintpp==0.4.0
preshed==3.0.9
prettytable==3.10.2
priority==2.0.0
proglog==0.1.10
protobuf==4.25.4
psutil==6.0.0
py-cpuinfo==9.0.0
py4j==0.10.9.7
py7zr==0.22.0
pyarrow==17.0.0
pybcj==1.0.2
pybind11==2.13.6
pycocoevalcap==1.2
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic-core==2.20.1
pydash==8.0.4
pydeck==0.9.1
pydub==0.25.1
pyenchant==3.2.2
pygments==2.18.0
pymongo==4.8.0
pynini==2.1.6
pynndescent==0.5.13
pyparsing==3.1.2
pypinyin==0.52.0
pyppmd==1.1.0
pysbd==0.3.4
pysocks==1.7.1
pyspark==3.5.1
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-wpe==0.0.1
pytz==2024.1
pyyaml==6.0.2
pyzmq==26.2.0
pyzstd==0.16.2
qwen-vl-utils==0.0.2
rarfile==4.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
retrying==1.3.4
rich==13.7.1
rpds-py==0.20.0
ruamel-yaml==0.18.6
ruamel-yaml-clib==0.2.8
ruff==0.5.7
s3transfer==0.10.4
sacrebleu==1.5.1
safehttpx==0.1.6
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.13.0
setproctitle==1.3.3
shellingham==1.5.4
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.4.0
sphn==0.1.4
spinners==0.0.24
srsly==2.4.8
sse-starlette==0.10.3
sseclient-py==1.8.0
st-multimodal-chatinput==0.2.1
starlette==0.40.0
strawberry-graphql==0.251.0
streamlit==1.39.0
streamlit-mic-recorder==0.0.8
sudachidict-core==20240716
sudachipy==0.6.8
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardx==2.6.2.2
termcolor==2.4.0
text-generation==0.7.0
texttable==1.7.0
thinc==8.2.5
threadpoolctl==3.5.0
thrift==0.16.0
tifffile==2024.8.10
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.20
toml==0.10.2
tomlkit==0.12.0
tornado==6.4.1
tqdm==4.66.5
transformer-engine==1.9.0
transformer-engine-cu12==1.9.0
transformers==4.45.0
transformers-stream-generator==0.0.5
triton==3.0.0
typeguard==4.3.0
typer==0.12.5
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
umap-learn==0.5.6
unidecode==1.3.8
universal-analytics-python3==1.1.1
urllib3==2.2.2
uvicorn==0.30.6
video2dataset==1.0.0
voxel51-eta==0.13.0
wandb==0.17.8
wasabi==1.1.3
watchdog==4.0.2
wcwidth==0.2.13
webdataset==0.2.100
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.8.0
websockets==12.0
werkzeug==3.0.3
wetextprocessing==1.0.4.1
wget==3.2
wrapt==1.16.0
wsproto==1.2.0
xmltodict==0.13.0
xxhash==3.4.1
yarg==0.1.9
yarl==1.9.4
yt-dlp==2024.8.6
zhconv==1.4.3
zhon==2.0.2
zipp==3.20.1
zope-event==5.0
zope-interface==7.0.3
